Begin transliteraterating diffs from 8.1 to 9.0.

+++ resources/src/main/java/org/robolectric/res/android/ByteBucketArray.java
@@ -20,7 +20,7 @@ public abstract class ByteBucketArray<T> {
     }
 
 //    byte bucketIndex = static_cast<byte>(index) >> 4;
-    byte bucketIndex = (byte) (((byte) index) >> 4);
+    byte bucketIndex = (byte) (index >> 4);
     T[] bucket = (T[]) mBuckets[bucketIndex];
     if (bucket == null) {
       return mDefault;
